+++ features/Layout/components/navbar/mobile-nav-item/MobileNavItem.tsx
@@ -1,17 +1,9 @@
 import { FunctionComponent } from 'react';
 
-import {
-  Flex,
-  Text,
-  Stack,
-  Collapse,
-  Icon,
-  Link,
-  useColorModeValue,
-  useDisclosure,
-} from '@chakra-ui/react';
+import { Flex, Text, Stack, Collapse, Icon, Link, useDisclosure } from '@chakra-ui/react';
 
-import { FaChevronDown } from 'react-icons/fa';
+import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
+import { faChevronDown } from '@fortawesome/free-solid-svg-icons';
 
 import { INavItem } from '../../../types';
 
@@ -24,33 +16,34 @@ const MobileNavItem: FunctionComponent<INavItem> = ({ label, children, href }) =
         py={2}
         as={Link}
         href={href ?? '#'}
-        justify={'space-between'}
-        align={'center'}
+        justify="space-between"
+        align="center"
         _hover={{
           textDecoration: 'none',
         }}
       >
-        <Text fontWeight={600} color={useColorModeValue('gray.600', 'gray.200')}>
+        <Text fontWeight={600} color="gray.600">
           {label}
         </Text>
         {children && (
           <Icon
-            as={FaChevronDown}
-            transition={'all .25s ease-in-out'}
+            as={FontAwesomeIcon}
+            icon={faChevronDown}
+            transition="all .25s ease-in-out"
             transform={isOpen ? 'rotate(180deg)' : ''}
             w={6}
             h={6}
           />
         )}
       </Flex>
-      <Collapse in={isOpen} animateOpacity style={{ marginTop: '0!important' }}>
+      <Collapse in={isOpen} animateOpacity style={{ marginTop: '0 !important' }}>
         <Stack
           mt={2}
           pl={4}
           borderLeft={1}
-          borderStyle={'solid'}
-          borderColor={useColorModeValue('gray.200', 'gray.700')}
-          align={'start'}
+          borderStyle="solid"
+          borderColor="gray.200"
+          align="start"
         >
           {children &&
             children.map((child) => (