+++ features/Layout/components/navbar/mobile-nav/MobileNav.tsx
@@ -2,9 +2,10 @@ import NextLink from 'next/link';
 
 import { FunctionComponent } from 'react';
 
-import { Stack, useColorModeValue } from '@chakra-ui/react';
+import { Stack, Icon } from '@chakra-ui/react';
 
-import { FaRss } from 'react-icons/fa';
+import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
+import { faRss } from '@fortawesome/free-solid-svg-icons';
 
 import { NAV_ITEMS } from 'features/Layout/constants';
 import { RSS_FEED_URL } from 'core/constants';
@@ -13,14 +14,12 @@ import MobileNavItem from '../mobile-nav-item/MobileNavItem';
 
 const MobileNav: FunctionComponent = () => {
   return (
-    <Stack bg={useColorModeValue('white', 'gray.800')} p={4} display={{ md: 'none' }}>
+    <Stack bg="white" p={4} display={{ md: 'none' }}>
       {NAV_ITEMS.map((navItem) => (
         <MobileNavItem key={navItem.label} {...navItem} />
       ))}
       <NextLink href={RSS_FEED_URL} passHref>
-        <a>
-          <FaRss cursor="pointer" />
-        </a>
+        <Icon as={FontAwesomeIcon} icon={faRss} cursor="pointer" />
       </NextLink>
     </Stack>
   );