+++ features/Layout/components/social-button/SocialButton.tsx
@@ -1,6 +1,6 @@
 import { FunctionComponent, ReactNode } from 'react';
 
-import { chakra, useColorModeValue, VisuallyHidden } from '@chakra-ui/react';
+import { chakra, VisuallyHidden } from '@chakra-ui/react';
 
 interface SocialButtonProps {
   children: ReactNode;
@@ -11,20 +11,20 @@ interface SocialButtonProps {
 const SocialButton: FunctionComponent<SocialButtonProps> = ({ children, label, href }) => {
   return (
     <chakra.button
-      bg={useColorModeValue('blackAlpha.200', 'whiteAlpha.100')}
+      bg="blackAlpha.200"
       rounded="full"
       w={8}
       h={8}
       cursor="pointer"
-      as={'a'}
+      as="a"
       href={href}
       display="inline-flex"
       alignItems="center"
       justifyContent="center"
-      transition="background 0.3s ease"
+      transition="background 0.4s ease"
       _hover={{
-        bg: useColorModeValue('blackAlpha.800', 'whiteAlpha.800'),
-        color: useColorModeValue('white', 'blackAlpha.800'),
+        bg: 'blackAlpha.800',
+        color: 'white',
       }}
     >
       <VisuallyHidden>{label}</VisuallyHidden>