+++ features/Metrics/index.ts
@@ -2,6 +2,6 @@ import { GA_TRACKING_ID } from './constants';
 
 import { GoogleAnalyticsScripts } from './components';
 
-import { gtagPageview } from './utilities/gtag';
+import { gtagPageview, gtagEvent } from './utilities/gtag';
 
-export { GA_TRACKING_ID, GoogleAnalyticsScripts, gtagPageview };
+export { GA_TRACKING_ID, GoogleAnalyticsScripts, gtagPageview, gtagEvent };