+++ features/Posts/components/posts-list/posts-list-item/PostsListItem.tsx
@@ -2,9 +2,10 @@ import { FunctionComponent } from 'react';
 
 import NextLink from 'next/link';
 
-import { FaChevronRight, FaClock } from 'react-icons/fa';
+import { Box, Flex, Heading, HStack, Tag, Text, Icon } from '@chakra-ui/react';
 
-import { Box, Flex, Heading, HStack, Tag, Text } from '@chakra-ui/react';
+import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
+import { faChevronRight, faClock } from '@fortawesome/free-solid-svg-icons';
 
 import { uppercaseFirst } from 'core';
 
@@ -43,9 +44,7 @@ const PostsListItem: PostsListItemType = ({
       <Flex marginBottom="1rem" align="center" justify="space-between">
         <Text>{`${date} - Written by ${authorName}`}</Text>
         <Tag backgroundColor="brand.greener" color="white">
-          <Box marginRight="0.5rem">
-            <FaClock />
-          </Box>
+          <Icon as={FontAwesomeIcon} icon={faClock} mr={2} />
           {readingTime}
         </Tag>
       </Flex>
@@ -59,7 +58,11 @@ const PostsListItem: PostsListItemType = ({
       {coverImage && <CoverImage coverImageSrc={coverImage} title={title} />}
       <Text marginBottom="1.25em">{excerpt}</Text>
       <Box marginBottom="1.5em">
-        <ButtonLink href={slug} label="Read More" rightIcon={<FaChevronRight />} />
+        <ButtonLink
+          href={slug}
+          label="Read More"
+          rightIcon={<Icon as={FontAwesomeIcon} icon={faChevronRight} />}
+        />
       </Box>
     </Flex>
   );